### 2025-07-30 9bc494f
require INLA in vignettes to pass CHECKS
