+++ DESCRIPTION
@@ -4,11 +4,14 @@ Version: 1.0.0
 Authors@R: c(
     person("Nathan", "Green", , "n.green@ucl.ac.uk", role = "aut",
            comment = c(ORCID = "0000-0003-2745-1736", ROR = "02jx3x895")),
-    person("Zhaojing", "Che", , "zhaojing.che.19@ucl.ac.uk", role = c("aut", "cph", "cre"),
+    person("Zhaojing", "Che", , "blendr-pkg@proton.me", role = c("aut", "cph", "cre"),
            comment = c(ORCID = "0000-0003-2245-1606", ROR = "052gg0110"))
   )
 Description: Create blended survival curves, see 'Che et al.' (2022) <doi:10.1177/0272989X221134545>.
 License: GPL (>= 3)
+Additional_repositories:
+    https://giabaio.r-universe.dev,
+    https://inla.r-inla-download.org/R/stable
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2.9000
@@ -25,9 +28,9 @@ Depends:
 Suggests:
     INLA,
     knitr,
+    remotes,
     rlang,
     rmarkdown,
-    remotes,
     survHEhmc,
     survival,
     testthat (>= 3.0.0)
