+++ cran-comments.md
@@ -1,13 +1,18 @@
 ## R CMD check results
 
 ❯ checking for future file timestamps ... NOTE
-  unable to verify current time
+unable to verify current time
 
 ❯ checking for unstated dependencies in vignettes ... NOTE
-  'library' or 'require' call not declared from: 'survHEhmc'
+'library' or 'require' call not declared from: 'survHEhmc'
 
 0 errors ✔ | 0 warnings ✔ | 2 notes ✖
 
 * This is a first release.
 
-* survHEhmc is not available on CRAN and so is Suggested. It is loaded conditionally in the vignettes.
+This package suggests the INLA and survHEhmc packages, which are
+available from the repositories listed in the Additional\_repositories field.
+
+The vignettes that depend on these packages have been pre-computed in inst/doc/
+to satisfy CRAN's build requirements.
+
