+++ cran-comments.md
@@ -1,16 +1,5 @@
 ## R CMD check results
 
-❯ checking for future file timestamps ... NOTE
-unable to verify current time
-
-0 errors ✔ | 0 warnings ✔ | 2 notes ✖
+0 errors ✔ | 0 warnings ✔ | 0 notes ✖
 
 * This is a first release.
-
-This package suggests the INLA and survHEhmc packages, which are
-available from the repositories listed in the Additional\_repositories field.
-
-The vignettes that depend on these packages have been pre-computed in inst/doc/
-to satisfy CRAN's build requirements.
-
->>>>>>> f46b4231e33a8f1644980aac9499c927627fb438
