+++ inst/paper.md
@@ -1,216 +1,0 @@
----
-title: 'blendR: An R package for blending survival curves'
-tags:
-  - R
-  - health economics
-  - survival analysis
-  - Bayesian inference
-authors:
-  - name: Nathan Green
-    orcid: 0000-0003-2745-1736
-    equal-contrib: true
-    affiliation: 1
-  - name: Zhaojing Che
-    orcid: 0000-0003-2745-1736
-    equal-contrib: true
-    affiliation: 2
-  - name: Gianluca Baio
-    corresponding: true
-    affiliation: 1
-affiliations:
-  - name: University College London (UCL), UK
-    index: 1
-    ror: 02jx3x895
-  - name: University of Oxford
-    index: 2
-    ror: 052gg0110
-date: 1 August 2025
-bibliography: paper.bib
-editor_options: 
-  markdown: 
-    wrap: 72
----
-
-# Summary
-
-The goal of `blendR` is to provide functionality to *blend* two survival
-curves together from one to the other according to some defined blending
-function.
-
-Survival extrapolation is essential in cost-effectiveness analysis to
-quantify the lifetime survival benefit associated with a new
-intervention, due to the restricted duration of randomized controlled
-trials (RCTs). Current approaches of extrapolation often assume that the
-treatment effect observed in the trial can continue indefinitely, which
-is unrealistic and may have a significant impact on decisions for resource
-allocation.
-
-Blended survival curves are a possible solution to alleviate the problem
-of survival extrapolation with heavily censored data from clinical
-trials. The main idea is to mix a flexible model (e.g., Cox
-semiparametric model) to fit as well as possible the observed data and a
-parametric model encoding assumptions on the expected behaviour of
-underlying long-term survival. The two are "blended" into a single
-survival curve that is identical with the first model over the range of
-observed times and gradually approaching the parametric model over the
-extrapolation period based on a weight function. The weight function
-regulates the way two survival curves are blended, determining how the
-internal and external sources contribute to the estimated survival over
-time. Long-term extrapolation from immature trial data may lead to
-significantly different estimates with various modelling assumptions.
-The blending approach provides sufficient flexibility, allowing a wide
-range of plausible scenarios to be considered as well as the inclusion
-of external information, based, for example, on hard data or expert
-opinion.
-
-# Statement of need
-
-Interim analyses of trials with limited follow-up are often subject to
-high degrees of administrative censoring, which may result in
-implausible long-term extrapolations using standard approaches [@NICE_TSD14].
-Implementing an innovative methodology based on "blending" survival
-curves to relax the traditional proportional hazard assumption and
-simultaneously incorporate external information can guide the
-extrapolation [@Che2022]. The `blendR` package provides a simple and
-powerful means to allow a careful consideration of a wide range of
-plausible scenarios, accounting for model fit to the short-term data as
-well as the plausibility of long-term extrapolations. `blendR` was
-designed to be used by statisticians, health economists, healthcare
-professionals and other users of survival data.
-
-# Method
-
-The *blending* idea is to consider two separate processes to describe
-the long-term horizon survival. The first one is driven exclusively by
-the observed data. Similar to a *standard* health technology assessments (HTA),
-this is used to
-determine an estimate over the entire time horizon of a survival curve,
-termed $S_{obs}(t \mid \boldsymbol{\theta}_{obs})$, a function of the
-relevant parameters $\boldsymbol{\theta}_{obs}$. A simple parametric
-model could be chosen, or alternatively, some other more complex model,
[…]
