+++ inst/paper.bib
@@ -1,44 +1,0 @@
-@article{Che2022,
-abstract = {Background Survival extrapolation is essential in the cost-effectiveness analysis to quantify the lifetime survival benefit associated with a new intervention, due to the restricted duration of randomized controlled trials (RCTs). Current approaches of extrapolation often assume that the treatment effect observed in the trial can continue indefinitely, which is unrealistic and may have a huge impact on decisions for resource allocation. Objective We introduce a novel methodology as a possible solution to alleviate the problem of performing survival extrapolation with heavily censored data from clinical trials. Method The main idea is to mix a flexible model (e.g., Cox semi-parametric) to fit as well as possible the observed data and a parametric model encoding assumptions on the expected behaviour of underlying long-term survival. The two are "blended" into a single survival curve that is identical with the Cox model over the range of observed times and gradually approaching the parametric model over the extrapolation period based on a weight function. The weight function regulates the way two survival curves are blended, determining how the internal and external sources contribute to the estimated survival over time. Results A 4-year follow-up RCT of rituximab in combination with fludarabine and cyclophosphamide v. fludarabine and cyclophosphamide alone for the first-line treatment of chronic lymphocytic leukemia is used to illustrate the method. Conclusion Long-term extrapolation from immature trial data may lead to significantly different estimates with various modelling assumptions. The blending approach provides sufficient flexibility, allowing a wide range of plausible scenarios to be considered as well as the inclusion of genuine external information, based e.g. on hard data or expert opinion. Both internal and external validity can be carefully examined.},
-archivePrefix = {arXiv},
-arxivId = {2206.00154},
-author = {Che, Zhaojing and Green, Nathan and Baio, Gianluca},
-eprint = {2206.00154},
-file = {:C\:/Users/Nathan/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Che, Green, Baio - 2022 - Blended Survival Curves A New Approach to Extrapolation for Time-to-Event Outcomes from Clinical Trial in H(2).pdf:pdf},
-month = {may},
-title = {{Blended Survival Curves: A New Approach to Extrapolation for Time-to-Event Outcomes from Clinical Trial in Health Technology Assessment}},
-url = {http://arxiv.org/abs/2206.00154},
-year = {2022}
-}
-
-@Article{survHE,
-  title = {{survHE}: Survival Analysis for Health Economic Evaluation and Cost-Effectiveness Modeling},
-  author = {Gianluca Baio},
-  journal = {Journal of Statistical Software},
-  year = {2020},
-  volume = {95},
-  number = {14},
-  pages = {1--47},
-  doi = {10.18637/jss.v095.i14},
-}
-
-@techreport{NICE_TSD14,
-  author = {Latimer, N. R. and Crowther, M. J. and Lambert, P. C. and Abrams, K. R. and Wailoo, A. J. and Morden, J. P. and Akehurst, R. L. and Campbell, M. J.},
-  title = {Survival analysis for economic evaluations alongside clinical trials -- extrapolation with patient-level data},
-  institution = {NICE Decision Support Unit},
-  year = {2013},
-  number = {TSD 14},
-  address = {Sheffield, UK},
-  url = {http://nicedsu.org.uk/technical-support-documents/survival-analysis-tsd/}
-}
-
-@Article{stan2017,
-  title = {{Stan}: A Probabilistic Programming Language},
-  author = {Bob Carpenter and Andrew Gelman and Matthew D. Hoffman and Daniel Lee and Ben Goodrich and Michael Betancourt and Marcus Brubaker and Jiqiang Guo and Peter Li and Allen Riddell},
[…]
