+++ BSCweb/www/note.md
@@ -1,1 +1,0 @@
-This folder includes the figures or other materials presented in the web.
