### 2025-08-28 #8672 @fauvernierma
Sorry but I am not available to perform this review

Have a great day

Sincerely,

Mathieu FAUVERNIER

Le mer. 27 août 2025 à 15:23, Oskar Laverny ***@***.***> a
écrit :

> *lrnv* left a comment (openjournals/joss-reviews#8672)
> <https://github.com/openjournals/joss-reviews/issues/8672#issuecomment-3228198087>
>
> @AleksandraDabrowska <https://github.com/AleksandraDabrowska>,
> @Nicolas-Schmidt <https://github.com/Nicolas-Schmidt>, @LingfengLuo0510
> <https://github.com/LingfengLuo0510>, @fauvernierma
> <https://github.com/fauvernierma>, would some of you be available, feel
> capable and be interested to review this submission for JOSS ? If you want
> details on what this implies, I can provide them.
>
