### 2025-08-29 977b3cd
addressed CRAN submission changes
