+++ cran-comments.md
@@ -1,5 +1,11 @@
 ## R CMD check results
 
-0 errors ✔ | 0 warnings ✔ | 0 notes ✖
+0 errors ✔ | 0 warnings ✔ | 1 notes ✖
+
+Regarding the NOTE about packages not in mainstream repositories:
+
+The dependencies INLA and survHEhmc are integral to my package's functionality. They are not available on CRAN but are correctly listed in the Additional_repositories field in the DESCRIPTION file, as confirmed by the check.
+
+Thank you for your consideration.
 
 * This is a first release.
