+++ DESCRIPTION
@@ -7,7 +7,16 @@ Authors@R: c(
     person("Zhaojing", "Che", , "blendr-pkg@proton.me", role = c("aut", "cph", "cre"),
            comment = c(ORCID = "0000-0003-2245-1606", ROR = "052gg0110"))
   )
-Description: Create blended survival curves, see 'Che et al.' (2022) <doi:10.1177/0272989X221134545>.
+Description: Create a blended curve from two survival curves, which is particularly useful
+          for survival extrapolation in health technology assessment. The main idea is to
+          mix a flexible model that fits the observed data well with a parametric model that
+          encodes assumptions about long-term survival. The two curves are blended into a
+          single survival curve that is identical to the first model over the range of 
+          observed times and gradually approaches the parametric model over the extrapolation
+          period based on a given weight function. This approach allows for the inclusion of
+          external information, such as data from registries or expert opinion, to guide
+          long-term extrapolations, especially when dealing with immature trial data.
+          See Che et al. (2022) <doi:10.1177/0272989X221134545>.
 License: GPL (>= 3)
 Additional_repositories: https://giabaio.r-universe.dev, https://inla.r-inla-download.org/R/stable
 Encoding: UTF-8
@@ -32,7 +41,6 @@ Suggests:
     survHEhmc,
     survival,
     testthat (>= 3.0.0)
-VignetteBuilder: knitr
 URL: https://github.com/StatisticsHealthEconomics/blendR/,
      https://StatisticsHealthEconomics.github.io/blendR/
 BugReports: https://github.com/StatisticsHealthEconomics/blendR/issues/
