### 2025-09-10 69cf42d
Revise installation instructions for blendR
Updated installation instructions for blendR.
