### 2025-09-22 #9055 @editorialbot
[REVIEW]: blendR: An R package for blending survival curves
**Submitting author:** @n8thangreen (<a href="http://orcid.org/0000-0003-2745-1736">Nathan Green</a>)
**Repository:** https://github.com/StatisticsHealthEconomics/blendR
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v1.0.0
