+++ README.md
@@ -1,5 +1,9 @@
 
+# blendR : An R package for blending survival curves <a href="https://StatisticsHealthEconomics.github.io/blendR/"><img src="man/figures/logo.png" align="right" height="139" alt="blendR website" /></a>
+  
+<!--
 # blendR: An R package for blending survival curves <img src='man/figures/hexbadge.png' style="float:right; height:200px;">
+-->
 
 <!-- badges: start -->
 [![R-CMD-check](https://github.com/StatisticsHealthEconomics/blendR/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/StatisticsHealthEconomics/blendR/actions/workflows/R-CMD-check.yaml)
