### 2025-09-29 2092f5f
improved help documentation
