+++ R/make_surv.R
@@ -2,39 +2,118 @@
 #' @title Create survival probabilities
 #' @name make_surv_methods
 #'
-#' @description These function are version of the [survHE::make.surv()] function
-#'    from \pkg{survHE}. These are needed prior to blending.
+#' @description
+#' A generic S3 function and methods to generate a standardized matrix of survival
+#' probabilities from various fitted survival model objects.
 #'
-#' @param Surv  survival analysis object
-#' @param ... Additional arguments
-#' @return Matrix of survival probabilities
+#' This function standardizes the output from different survival modelling packages
+#' into a consistent format: a matrix where rows represent discrete time points
+#' and columns represent simulations from the model's posterior distribution. This
+#' standardized format is essential for use in downstream evidence blending
+#' functions. The methods are inspired by the [survHE::make.surv()] function.
+#'
+#' @param Surv A fitted survival model object or a matrix/vector of survival
+#'   probabilities. Supported classes include `survHE`, `flexsurvreg`, `inla`,
+#'   `matrix`, and `numeric`.
+#' @param ... Additional arguments passed to specific methods (e.g., `t`, `nsim`).
+#' @return A matrix with `length(t)` rows and `nsim` columns. Each element `[i, j]`
+#'   is the survival probability at time `t[i]` for simulation `j`.
+#'
+#' @seealso [survHE::make.surv()]
+#'
 #' @export
 #'
-#' @examplesIf rlang::is_installed("survHEhmc")
-#' library(survHE)
+#' @examples
+#' # Define common time points and number of simulations for examples
+#' time_points <- 1:100
+#' n_sim <- 50
 #'
-#' ## trial data
-#' data("TA174_FCR", package = "blendR")
+#' #--------------------------------------
+#' ## Method for a 'survHE' object
+#' #--------------------------------------
+#' if (rlang::is_installed("survHE") && rlang::is_installed("survival")) {
+#'   library(survHE)
+#'   library(survival)
+#'   data(ovarian)
 #'
-#' ## externally estimated data
-#' data_sim <- ext_surv_sim(t_info = 144,
-#'                          S_info = 0.05,
-#'                          T_max = 180)
+#'   # Fit a Weibull model using survHE (with MLE for speed)
+#'   fit_she <- fit.models(
+#'     formula = Surv(futime, fustat) ~ 1,
+#'     data = ovarian,
+#'     distr = "weibull",
+#'     method = "mle"
+#'   )
 #'
-#' ext_Surv <- fit.models(formula = Surv(time, event) ~ 1,
-#'                        data = data_sim,
-#'                        distr = "exponential",
-#'                        method = "hmc")
+#'   # Generate survival probability matrix
+#'   surv_matrix_she <- make_surv(fit_she, t = time_points, nsim = n_sim)
+#'   cat("survHE method output dimensions:", dim(surv_matrix_she), "\n")
+#' }
 #'
-#' S_ext <- make_surv(ext_Surv, t = 1:100, nsim = 100)
+#' #--------------------------------------
+#' ## Method for a 'flexsurvreg' object
+#' #--------------------------------------
+#' if (rlang::is_installed("flexsurv") && rlang::is_installed("survival")) {
+#'   library(flexsurv)
+#'   library(survival)
 #'
+#'   # Fit a log-logistic model using flexsurv
+#'   fit_fsr <- flexsurvreg(
+#'     formula = Surv(futime, fustat) ~ 1,
+#'     data = ovarian,
+#'     dist = "llogis"
+#'   )
+#'
+#'   # Generate survival probability matrix
[…]
