+++ R/ext_surv_sim.R
@@ -1,22 +1,45 @@
 
 #' Create an external survival data based on expert opinion
 #'
-#' Generally, the sampling is done is two steps
+#' @description
+#' Generates an individual patient-level survival dataset from aggregate survival
+#' probabilities that might be elicited from expert opinion.
+#'
+#' The function creates a synthetic dataset of event times for `n` patients,
+#' consistent with a piecewise-constant hazard rate implied by the expert-provided
+#' survival probabilities.
+#'
+#' @details
+#' The simulation uses a two-step sampling process based on partitioning the time
+#' horizon from `0` to `T_max`. First, the total number of patients `n` is
+#' allocated to different time intervals `[t_i, t_{i+1}]`. The probability of an
+#' event occurring in an interval is derived from the drop in the survival curve
+#' over that interval. This allocation is done via a multinomial distribution:
 #' \deqn{
-#'    p(T) = p(T | interval i) p(interval i)
+#'     i \sim \text{Multinomial}(\pi)
+#' }
+#' where $\pi_i = S(t_{i-1}) - S(t_i)$ is the probability of an event in interval `i`.
+#'
+#' Second, for the patients assigned to each interval, a specific event time is
+#' simulated from a uniform distribution covering that interval:
+#' \deqn{
+#'     T | i \sim U(t_{i-1}, t_i)
 #' }
 #'
-#' In particular
-#' \eqn{T ~ U(x_{i}, x_{i+1})}
-#' \eqn{i ~ multinomial(\hat{\pi})}
-#'
-#' @param t_info A vector of times for which expert opinion is elicited
-#' @param S_info A vector of mean survival probabilities estimated by experts
-#'               corresponding to time points in `t_info`
-#' @param T_max  The maximum survival time to be used
-#' @param n      The number of patients to construct the artificial external data set; default 100
+#' @param t_info A numeric vector of time points for which expert opinion is elicited.
+#' @param S_info A numeric vector of mean survival probabilities estimated by experts
+#'               corresponding to time points in `t_info`.
+#' @param T_max  The maximum survival time for the simulation, at which the
+#'   survival probability is assumed to be 0.
+#' @param n      The total number of patients to construct the artificial external data set; default 100
 #' @importFrom stats runif rmultinom
-#' @return Dataframe of times and censoring status
+#'
+#' @return A data frame with `n` rows and two columns:
+#'   \itemize{
+#'     \item **time**: The simulated event time for each patient.
+#'     \item **event**: The event indicator, which is always `1` as this function does not simulate censoring.
+#'   }
+#'
 #' @export
 #'
 #' @examples
@@ -24,6 +47,7 @@
 #'                     S_info = c(0.9, 0.8, 0.2),
 #'                     T_max = 100, n = 100)
 #' if (require(survival)) {
+#'     # Kaplan-Meier curve
 #'     km_fit <- survfit(Surv(time, event) ~ 1, data = dat)
 #'     plot(km_fit)
 #' }

[…]
