### 2025-10-15 StatisticsHealthEconomics/blendR#12 @lrnv
[Readme] put badges and links up top
It would be nice if the different badges (r-cmd-check, licencing, link to the pkgdown docs website) would be at the top of the readme.md, which is their standard location.

Crosslink to https://github.com/openjournals/joss-reviews/issues/9055
