### 2025-10-16 3775619
Restore blendR package description and badge
Updated README to restore package description and badge.
