+++ README.md
@@ -1,6 +1,3 @@
-
-# blendR : An R package for blending survival curves <a href="https://StatisticsHealthEconomics.github.io/blendR/"><img src="man/figures/hexbadge.png" align="right" height="139" alt="blendR website" /></a>
-  
 <!--
 # blendR: An R package for blending survival curves <img src='man/figures/hexbadge.png' style="float:right; height:200px;">
 -->
@@ -9,6 +6,8 @@
 [![R-CMD-check](https://github.com/StatisticsHealthEconomics/blendR/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/StatisticsHealthEconomics/blendR/actions/workflows/R-CMD-check.yaml)
 <!-- badges: end -->
 
+# blendR : An R package for blending survival curves <a href="https://StatisticsHealthEconomics.github.io/blendR/"><img src="man/figures/hexbadge.png" align="right" height="139" alt="blendR website" /></a>
+  
 The goal of blendR is to _blend_ two survival curves together from one to the other according to some defined blending function.
 Originally developed for the following paper
 
