### 2025-10-16 StatisticsHealthEconomics/blendR#12 @giabaio
Is that what you had in mind, @lrnv (see current version of the `README` file)?
