### 2025-11-03 #9055 @lrnv
👋 @Nicolas-Schmidt , @NicolasNgo how is this rview going ? Please tell me if you need some help and or guidance to go through the content :)
