### 2025-12-18 #9055 @NicolasNgo
Hi @n8thangreen,

Overall, the blendR package provides a useful implementation of an already validated method for blending survival curves. The implementation is efficient and appropriate. The installation process is well described, as is the blending methodology in the submitted paper. However, I have a few suggestions and comments with regarding the documentation.

**General comments**
- Add the CRAN badge on top of the README page.
- The CRAN documentation differs, in some cases, from the R documentation (e.g., `ext_surv_sim`).

**blendsurv**
- Add references to the underlying method in the `blendsurv` function description, ideally in a References section.
- Provide the exact names of the functions and objects that can be use for `obs_Surv` and `ext_Surv`. An exhaustive list of the accepted objects could be added in a Details section.
- Related to the previous point, I suggest adding a link to the R documentation of the functions that be can use for the observed and external survival fits. These could be included in a See also section.
- Explicitly describe the required format of the `beta_params` argument (e.g., does it have to be a named list with `alpha` and `beta`, or can it be a vector ?).
- In `beta_params` description, use a capital 'C' for _coefficients_.
- The documentation would benefit from a more complete description of the output of the `blendsurv` function.
- The `blend_interv` argument corresponds to the [a, b] interval described in the submitted paper. Using the same notations instead of `min` and `max ` may help avoid confusion.
- I do not understand the purpose of the `nsim` argument. At which step of the blending process is simulation required?
- The package `survHEhmc` is required to run the example but is not explicitely mentioned. I had difficulty installing it and had to rely on the dedicated `survHEhmc` documentation to install it properly and run your example. It would be helpful to include this information on how to install `survHEhmc` properly.
- It could be interesting to include an additional example using a different observed survival fit.

**dat_FCR**
- Use consistent naming between the shared data in the example and in the documentation (dat_FCR vs TA174_FCR).
- The shared dataset would benefit a short description explaining the values, context and how the data were collected.

**ext_surv_sim**
For this function, I mainly relied on the R documentation because the CRAN documentation is largely missing or incorrect.
- In the R documentation, some mathematical notations are not rendered properly (e.g., [t_i, t_{i+1}] or $\pi_i = S(t_i-1) - S(t_i)$).
- The notation i ~ Multinomial(π) is unclear, as the multinomial distribution requires two parameters. It is also unclear whether the goal is to randomly select the indices at which the events will occur (how many ?), or whether there is a missing notation and the number of events $n_{i}$ in the interval $i$ is being randomly drawn.
- Perhaps $\pi_i$ is intended to be an argument of the multinomial distribution rather than $\pi$ alone.

**fit_inla_pw**
- Missing argument description for `nsim`.

**manip_plot**
- It is unclear what `manip_plot` function does and how it should be used.
- I was unable to run the code for this function

**plot**
- The warning '_Using `size` aesthetic for lines was deprecated in ggplot2 3.4.0. Please use `linewidth` instead_' appears. Updating this would avoid unnecessary warnings.
- It could be interesting to include an argument allowing the user to plot a single curve instead of all three survival curves. This is only a suggestion.

**weightplot**
- The function runs correctly, but I obtained the following warning using it with the `ble_Surv` example object:
‘_All aesthetics have length 1, but the data has 201 rows.
ℹ Please consider using `annotate()` or provide this layer with data containing a single row._’

**Paper**
- Missing '_do_' at line 73 '_generally desirable to so._'
- '_Figure Figure 1_' repeated (line 64 and 78).
- I assume the shaded areas around the survival curves represent confidence intervals  for the blended survival curves. It would be helpful to explain how these intervals are estimated (or simply refered to the paper explaining the procedure).
- I am simply curious whether it is possible to use alternative weight functions instead of the cumulative distribution function of a Beta random variable.

Most of the comments above relate to documentation and user guidance, with the goal of making the package easier to understand and use for new  users. The blendR package remains a valuable and well-implemented tool based on a validated approach to survival curve blending.
