### 2026-09-01 #9055 @n8thangreen
https://github.com/openjournals/joss-reviews/issues/9055#issuecomment-3669892156
CRAN badge added Commit 97213a1
97213a198698ef52e382dac083188e5238c8a261
