### 2026-09-01 #9055 @n8thangreen
> Thanks [@Nicolas-Schmidt](https://github.com/Nicolas-Schmidt) ! [@n8thangreen](https://github.com/n8thangreen) can you please review the paper itself with respect to these comments ?

I've addressed them here
ebc418b0ec8ca8f40c44de0f72aba976a1abb92c
