### 2026-09-01 #9055 @n8thangreen
> * Add references to the underlying method in the `blendsurv` function description, ideally in a References section.
Done as requested:
4b62c516dd91396a79b3b7e6a1dd5f94d55409c7
