### 2026-09-01 #9055 @n8thangreen
> * Provide the exact names of the functions and objects that can be use for `obs_Surv` and `ext_Surv`. An exhaustive list of the accepted objects could be added in a Details section.

Addressed with 9704ca4205ac6f427081d4aee0e8f315f1a457e1
