### 2026-09-01 #9055 @n8thangreen
> * Related to the previous point, I suggest adding a link to the R documentation of the functions that be can use for the observed and external survival fits. These could be included in a See also section.

Addressed 155a0f07e9eb70ff9727d2f152b981abedf61cd4
