### 2026-09-01 #9055 @n8thangreen
> * Explicitly describe the required format of the `beta_params` argument (e.g., does it have to be a named list with `alpha` and `beta`, or can it be a vector ?).

Done
91853b338c36bbd1b0316b0f1b2fbaea3a11e739
