### 2026-09-01 #9055 @n8thangreen
> * The documentation would benefit from a more complete description of the output of the `blendsurv` function.

Done
1446c6ded6cad97e1bc03337ae2bb15096901229
