### 2026-09-01 #9055 @n8thangreen
> * I do not understand the purpose of the `nsim` argument. At which step of the blending process is simulation required?

Improved details:
2908688880a92af6a790fb8ff8e8ca2381ffb299
