### 2026-09-03 #9055 @lrnv
@n8thangreen Could you please give us a summary of which part of the reviews were already taken into account / which ones are still on the table ?
