### 2026-09-03 #9055 @n8thangreen
> * Use consistent naming between the shared data in the example and in the documentation (dat_FCR vs TA174_FCR).

addresses: 650c4c28a2b8c372b2350413497e9c03fe3cba88
