### 2026-09-03 #9055 @n8thangreen
> **ext_surv_sim** For this function, I mainly relied on the R documentation because the CRAN documentation is largely missing or incorrect.

This is a good pick up. I think I meant to go back to this. Now fixed.
00a010e9788d5a099c7fa7473710deb46102ceec
