### 2026-09-03 #9055 @n8thangreen
> [@n8thangreen](https://github.com/n8thangreen) Could you please give us a summary of which part of the reviews were already taken into account / which ones are still on the table ?

Hi, I've got to fix the `weightplot()` warning and the paper comments.
